+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `Innmind\HttpSession\Session::maybe()`
+
 ### Changed
 
 - Require php 8.1